+++ tests/test_edf.py
@@ -5,6 +5,7 @@ import io
 import json
 import tempfile
 from pathlib import Path
+from shutil import copyfile
 from typing import Literal
 
 import numpy as np
@@ -1160,3 +1161,39 @@ def test_sampling_frequencies_leading_to_floating_point_issues_in_signal_duratio
     assert edf.num_data_records == 10
     assert edf.signals[0].samples_per_data_record == 22
     assert edf.signals[1].samples_per_data_record == 9
+
+
+# fmt: off
+@pytest.mark.parametrize(
+    ("extra_bytes", "num_records_in_header", "expected_warning"),
+    [
+    #    extra bytes     num records field    expected warning
+        (1,              10,                  "Incomplete data record at the end of the EDF file"),
+        (15,             11,                  "Incomplete data record at the end of the EDF file"),
+        (0,              9,                   "EDF header indicates 9 data records, but file contains 10 records. Updating header."),
+        (0,              11,                  "EDF header indicates 11 data records, but file contains 10 records. Updating header."),
+    ],
+)
+# fmt: on
+def test_read_edf_with_invalid_number_of_records(
+    tmp_path: Path,
+    extra_bytes: int,
+    num_records_in_header: int,
+    expected_warning: str,
+):
+    invalid_edf = tmp_path / "invalid.edf"
+    copyfile(EDF_FILE, invalid_edf)
+    with invalid_edf.open("rb+") as file:
+        file.seek(236)
+        file.write(f"{num_records_in_header: <8}".encode("ascii"))
+        if extra_bytes > 0:
+            file.seek(0, 2)
+            file.write(b"\0" * extra_bytes)
+
+    for io_obj in (invalid_edf, invalid_edf.read_bytes()):
+        with pytest.warns(UserWarning, match=expected_warning):
+            edf = read_edf(io_obj)
+        assert edf.num_data_records == 10
+        comparison_edf = read_edf(EDF_FILE)
+        for signal, comparison_signal in zip(edf.signals, comparison_edf.signals):
+            np.testing.assert_array_equal(signal.data, comparison_signal.data)